Add stable support
Tested with: `cargo +stable test`
