+++ src/lib.rs
@@ -1,45 +1,4 @@
-#![forbid(unsafe_code)]
+use proc_macro_hack::proc_macro_hack;
 
-extern crate proc_macro;
-
-use proc_macro::TokenStream;
-use syn::parse::Parse;
-use syn::parse::ParseStream;
-use syn::parse::Result;
-use syn::{Token, parse_macro_input};
-
-use quote::quote;
-use syn::Expr;
-use syn::LitInt;
-
-struct ArrayInit {
-    value: Expr,
-    quantity: LitInt,
-}
-
-impl Parse for ArrayInit {
-    fn parse(input: ParseStream) -> Result<Self> {
-        let value: Expr = input.parse()?;
-        input.parse::<Token![;]>()?;
-        let quantity: LitInt = input.parse()?;
-        Ok(ArrayInit {
-            value,
-            quantity,
-        })
-    }
-}
-
-#[proc_macro]
-pub fn arr(input: TokenStream) -> TokenStream {
-    let ArrayInit {
-        value,
-        quantity
-    } = parse_macro_input!(input as ArrayInit);
-
-    let iter = std::iter::repeat(value).take(quantity.value() as usize);
-
-    let result = quote! {
-        [#(#iter),*]
-    };
-    result.into()
-}
+#[proc_macro_hack]
+pub use arr_macro_impl::arr;
